+++ test/noahtheduke/splint/rules/lint/try_splicing_test.clj
@@ -1,0 +1,12 @@
+; This Source Code Form is subject to the terms of the Mozilla Public
+; License, v. 2.0. If a copy of the MPL was not distributed with this
+; file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+(ns noahtheduke.splint.rules.lint.try-splicing-test
+  (:require
+    [expectations.clojure.test :refer [defexpect expect]]
+    [noahtheduke.splint-test :refer [check-alt]]))
+
+(defexpect try-splicing-test
+  (expect '(try (do (splint/unquote-splicing body)) (finally :true))
+    (check-alt "(try ~@body (finally :true))")))